app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "0.16.0"
fastapi = "*"
fastapi-utils = "^0.2.1"
pydrive = "^1.3.1"
fpdf2 = "^2.6.1"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
celery = "^4.4.2"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
pydantic = "^1.4"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.2"
aiofiles = "0.7.0"
psycopg2-binary = "^2.8.5"
importlib-metadata="4.13.0"
alembic = "^1.4.2"
sqlalchemy = "^1.3.16"
pytest = "^5.4.1"
unidecode = "1.3.2"
XlsxWriter = "3.0.2"
Openpyxl = "3.0.9"
websockets = "^10.4"
apscheduler = "*"
numpy = "*"
watchfiles = "0.18.1"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
pymysql = "*"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"
qrcode = "^7.3.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

